+++ src/PSRule/Commands/SetPSRuleHintCommand.cs
@@ -12,9 +12,6 @@ namespace PSRule.Commands
         [Parameter(Mandatory = false, Position = 0)]
         public string Message { get; set; }
 
-        [Parameter(Mandatory = false)]
-        public string TargetName { get; set; }
-
         protected override void ProcessRecord()
         {
             var result = GetResult();
@@ -22,11 +19,6 @@ namespace PSRule.Commands
             if (MyInvocation.BoundParameters.ContainsKey("Message"))
             {
                 result.Message = Message;
-            }
-
-            if (MyInvocation.BoundParameters.ContainsKey("TargetName"))
-            {
-                PipelineContext.CurrentThread.WarnTargetNameParameterObsolete();
             }
         }
     }